core/requirements.txt
Django>=3.0,<4.0
djangorestframework-simplejwt
djangorestframework
django-cors-headers
djangorestframework